Changing FileHolder network design
* Adding FileServer concept so devices can pass a persistent reference to the remote file system for file transfer.
* FileHolder is still a network object, but now servants are created locally during transfer instead of at file creation.
* Adding VirtualFileHolder to allow file data to stay in memory before transfer instead of on disk.
* Adding unique FileID to identify files.
* Fixed issue with stipy saving timing files (was saving jupyter internal python files).
* Timing files now save in correct relative directory when shot is saved.

Jason

+++ src/network/src/TTaskManager_i.cpp
@@ -136,7 +136,7 @@ void TTaskManager_i::runTask(const char* taskID)
     }
 }
 
-::CORBA::Float TTaskManager_i::secondsToNextRun(const char* taskID)
+::CORBA::Double TTaskManager_i::secondsToNextRun(const char* taskID)
 {
     std::shared_ptr<Task> task;
     double result = 0;
@@ -144,7 +144,7 @@ void TTaskManager_i::runTask(const char* taskID)
     if (taskManager != 0 && taskManager->getTask(taskID, task) && task != 0) {
         result = task->secondsToNextRun();
     }
-    return static_cast<::CORBA::Float>(result);
+    return static_cast<::CORBA::Double>(result);
 }
 
 ::CORBA::Boolean TTaskManager_i::ping()
