+++ src/network/src/TTaskManager_i.h
@@ -31,7 +31,7 @@ public:
     void activateTask(const char* taskID);
     void deactivateTask(const char* taskID);
     void runTask(const char* taskID);
-    ::CORBA::Float secondsToNextRun(const char* taskID);
+    ::CORBA::Double secondsToNextRun(const char* taskID);
     ::CORBA::Boolean ping();
 
 private:
